Fix issue #3: Do not sleep 2 sec before exiting main loop.
The `sigquit_cb()` did not properly exit the libuEv event loop, instead
it relied on it receiving `SIGKILL`, which it sent itself after sleeping
for a couple of seconds.  The intent was to let any FTP/TFTP sub-process
handle SIGTERM gracefully.

This patch changes the `sigquit_cb()` to only forward `SIGTERM` to its
process group, `sched_yield()` instead of `sleep(2)`, and then properly
call `uev_exit()`.

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>