Refactor FTP and TFTP sessions to leverage libuEv better
This is a massive refactor.  It started out with finding a use-case
where the CPU load suddenly went through the roof!   After a bit of
investigating I realised I had not yet fully converted uftpd to the
libuEv library.  So this patch is actually the continuation of that
work.  In the process I also cleaned up the FTP code a bit and also
added a few new FTP functions: HELP and FEAT, used by some clients.

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>