Leverage in.tftpd and in.ftpd symlinks for inetd usage
This patch removes the `-t PORT`, and `-f PORT` arguments, as well as
`-i` to enable inetd mode.  To use uftpd from inetd the new symlinks
in.tftpd and in.ftpd have to be used.

In inetd mode both foreground and syslog is now enforced.

The ability to run uftpd TFTP or FTP on different ports will be restored
in a later commit using `-o tftp=PORT` and `-o ftp=PORT`, using POSIX
getsubopt().

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>