Update build system to re-use configure run from parent dir in libs
This patch updates the uftpd build system to generate the Makefiles
for both libuev and libite, now that both have been converted to the
GNU configure and build system.  This means we only need to run the
configure script once, but since automake et al are a bit stupid it
wants to call the `install` target in the submodules, which we do
not want.

Also add check for missing GIT submodules.  We still need to add
`--with-libuev=""` and `--with-libite=""` switches to configure so
the user can use .so files from the system.

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>