Refactor, simplify recv_msg(): no need for while() loop anymore
The command receive callback no longer needs to use a while() loop,
i.e. to do continue on EAGAIN.  This is handled by the event loop.

Also, recv() may return ECONNRESET when a client simply drops the
connection with close().  We don't need to call foul on that, so
hide that in a DBG() message instead.

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>