Add limited support for MLST OPTS, fact ordering
The default MLST fact ordering in uftpd is "modify;perm;size;type;", but
a client may send "OPTS MLST perm;type;size;modify;" to change the order
in which we list facts.  This patch adds limited support for that, it is
limited in the sense that uftpd is still limited to these four facts.

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>